manager panel component updated
