--- client/src/app/components/manageer-pannel/manageer-pannel.component.scss
@@ -1,268 +1,0 @@
-#main-section {
-    width: 100%;
-    margin-top: 8rem;
-
-    .register-form {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin: 2rem 0;
-
-        .content {
-            width: 90%;
-            display: flex;
-            flex-direction: column;
-            flex-wrap: wrap;
-            gap: 2rem;
-            align-items: center;
-            padding: 1.5rem 1rem;
-
-            border-radius: 10px;
-            background: linear-gradient(145deg, #f0f0f0, #cacaca);
-            box-shadow: 5px 5px 10px #5a5a5a, -5px -5px 10px #ffffff;
-
-            .div-title-radio-input {
-                flex: 0 0 100%;
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                align-items: center;
-                width: 100%;
-                margin-top: .5rem;
-                gap: 1rem;
-
-                .title {
-                    font-size: 20px;
-                    color: #000;
-                    font-weight: 800;
-                }
-
-                .radio-group {
-                    display: flex;
-                    justify-content: center;
-                    align-items: center;
-                    gap: 1rem;
-
-                    .radio-button {
-                        height: 60px;
-                        width: 100px;
-                        display: flex;
-                        justify-content: center;
-                        align-items: center;
-                        border-radius: 10px;
-                        background: linear-gradient(145deg, #f0f0f0, #cacaca);
-                        box-shadow: 5px 5px 10px #5a5a5a, -5px -5px 10px #ffffff;
-                    }
-                }
-            }
-
-            .form-item {
-                width: 90%;
-                max-width: 700px;
-                display: flex;
-                flex-direction: column;
-                align-items: stretch;
-                min-width: 0;
-
-                mat-form-field {
-                    display: block;
-                    width: 100%;
-
-                    .input {
-                        padding: 10px;
-                        height: 54px;
-                        width: 100%;
-                        border: none;
-                        border-radius: 10px;
-                        background: transparent;
-                        box-shadow: none;
-                    }
-
-                    mat-label {
-                        font-size: 15px;
-                        font-weight: 800;
-                        color: #000;
-                    }
-
-                    mat-hint {
-                        font-size: 12px;
-                    }
-
-                    mat-error {
-                        color: #d32f2f;
-                        font-size: 12px;
-                        line-height: 1.6;
-                        display: block;
-                    }
-                }
-
-                .field-error {
-                    margin-top: 6px;
-                    padding: 0 4px;
-                    font-size: 12px;
-                    color: #d32f2f;
-                    line-height: 1.6;
-                    display: flex;
-                    flex-direction: column;
-                    gap: 2px;
-                    width: 100%;
-                }
-            }
-
-            .div-btn-add {
-                flex: 0 0 100%;
-                width: 100%;
-                display: flex;
-                justify-content: center;
-                align-items: center;
-                margin: .5rem 0 1rem;
-
-                button {
-                    border-radius: 10px;
-                    border: none;
-                    height: 40px;
-                    width: 120px;
-                }
-            }
-        }
-    }
-}
-
-@media (min-width: 768px) {
-    #main-section .register-form .content {
-        width: 90%;
-        flex-direction: row;
-        justify-content: center;
-        align-items: stretch;
-    }
-
-    #main-section .register-form .content .form-item {
-        flex: 1 1 45%;
-        max-width: 45%;
-        width: auto;
-    }
-}
-
-@media (min-width: 1200px) {
-    #main-section {
-        margin-top: 0;
-    }
-
-    #main-section .register-form .content {
-        width: 60%;
-        justify-content: space-between;
-    }
-
-    #main-section .register-form .content .form-item {
-        flex: 1 1 30%;
-        max-width: 30%;
-    }
-}
-
-::ng-deep mat-tab-group {
-    margin-top: 6rem;
-}
-
-::ng-deep mat-form-field {
-    height: auto !important;
-    margin-bottom: 10px;
-}
-
-::ng-deep .mat-mdc-text-field-wrapper {
-    border-radius: 12px;
-    background: #e0e0e0;
-    box-shadow: inset 5px 5px 10px #cecece, inset -5px -5px 10px #f2f2f2;
-    padding: 0;
-    overflow: visible;
-}
-
-::ng-deep .mat-mdc-form-field-appearance-outline .mdc-notched-outline {
-    display: none;
-}
-
-::ng-deep .mat-mdc-form-field-infix {
-    text-align: right;
-    direction: rtl;
-    padding: 12px 14px !important;
-    min-height: 54px;
-    width: 100%;
-    box-sizing: border-box;
-}
-
-::ng-deep .mdc-floating-label {
-    white-space: nowrap;
-    overflow: visible;
-    text-overflow: clip;
-}
-
-::ng-deep .mat-mdc-form-field-subscript-wrapper {
-    position: relative;
-    bottom: 0;
-    margin-top: 6px;
-    padding: 0 12px;
-    height: auto;
-}
-
-::ng-deep .mdc-text-field--invalid .mdc-notched-outline__leading,
-::ng-deep .mdc-text-field--invalid .mdc-notched-outline__notch,
-::ng-deep .mdc-text-field--invalid .mdc-notched-outline__trailing {
-    border-color: transparent !important;
-}
-
-::ng-deep .mdc-text-field--invalid .mat-mdc-text-field-wrapper {
-    border: none !important;
-}
-
-::ng-deep mat-form-field.phone-field .mat-mdc-form-field-prefix {
-    direction: ltr;
-    display: flex;
-    align-items: center;
-    height: 100%;
-    padding-inline-start: 4px;
-    padding-inline-end: 4px;
-    font-weight: 700;
-    opacity: .9;
-}
-
-::ng-deep mat-form-field.phone-field input.mat-mdc-input-element {
-    line-height: 1.5;
-    text-align: left;
-}
-
-::ng-deep mat-form-field.phone-field .mat-mdc-text-field-wrapper {
-    padding: 0 12px;
-}
-
-::ng-deep mat-form-field.phone-field .phone-label {
-    margin-left: 10px;
-}
-
-.field-error {
-    direction: rtl;
-    text-align: right;
-    white-space: pre-line;
-    margin-top: 6px;
-    padding: 0 4px;
-    font-size: 12px;
-    color: #d32f2f;
-    line-height: 1.8;
-}
-
-.field-error span {
-    display: block;
-}
-
-::ng-deep .mat-mdc-form-field-subscript-wrapper,
-::ng-deep .mat-mdc-form-field-hint,
-::ng-deep .mat-mdc-form-field-error {
-    direction: rtl !important;
-    text-align: right !important;
-    justify-content: flex-end;
-    white-space: pre-line;
-}
-
-::ng-deep .mat-mdc-form-field-bottom-align::before,
-::ng-deep .mat-mdc-form-field-bottom-align::after {
-    content: none;
-}
