+++ client/src/app/components/manager-panel/manager-panel.component.html
@@ -1,0 +1,21 @@
+<header>
+  <nav>
+    <app-navbar />
+  </nav>
+</header>
+
+<section id="main-section" class="panel">
+  <mat-tab-group>
+    <mat-tab label="ثبت نام منشی">
+      <app-register-secretary />
+    </mat-tab>
+
+    <mat-tab label="ثبت نام دانشجو">
+      <app-register-student />
+    </mat-tab>
+
+    <mat-tab label="ثبت نام مدرس">
+      <app-register-teacher />
+    </mat-tab>
+  </mat-tab-group>
+</section>
