+++ client/src/app/components/register/register-secretary/register-secretary.component.html
@@ -1,0 +1,122 @@
+<form class="register-form" [formGroup]="addSecretaryFg" (ngSubmit)="addSecretary()" #secForm="ngForm">
+    <div class="content">
+        <div class="div-title-radio-input">
+            <span class="title">ثبت نام منشی</span>
+            <mat-radio-group class="radio-group" aria-label="Select an option" formControlName="genderCtrl">
+                <mat-radio-button class="radio-button" value="male">مرد</mat-radio-button>
+                <mat-radio-button class="radio-button" value="female">زن</mat-radio-button>
+            </mat-radio-group>
+        </div>
+
+        <!-- name -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>نام</mat-label>
+                <input class="input" type="text" formControlName="nameCtrl" matInput dir="auto">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryNameCtrl)">
+                <span *ngIf="SecretaryNameCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryNameCtrl.hasError('minlength')">خیلی کوتاهه حداقل: 2 حرف</span>
+                <span *ngIf="SecretaryNameCtrl.hasError('maxlength')">خیلی بلنده حداکثر: 30 حرف</span>
+                <span *ngIf="SecretaryNameCtrl.hasError('pattern')">فقط حروف فارسی/انگلیسی، فاصله و خط‌تیره مجاز
+                    است.</span>
+            </div>
+        </div>
+
+        <!-- lastName -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>نام خانوادگی</mat-label>
+                <input class="input" type="text" formControlName="lastNameCtrl" matInput dir="auto">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryLastNameCtrl)">
+                <span *ngIf="SecretaryLastNameCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryLastNameCtrl.hasError('minlength')">خیلی کوتاهه حداقل: 2 حرف</span>
+                <span *ngIf="SecretaryLastNameCtrl.hasError('maxlength')">خیلی بلنده حداکثر: 30 حرف</span>
+                <span *ngIf="SecretaryLastNameCtrl.hasError('pattern')">فقط حروف فارسی/انگلیسی، فاصله و خط‌تیره مجاز
+                    است.</span>
+            </div>
+        </div>
+
+        <!-- email -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>ایمیل</mat-label>
+                <input class="input" autoFocusDir type="email" formControlName="emailCtrl" matInput maxlength="50"
+                    dir="auto">
+                <mat-hint>ایمیل خودرا وارد کنید</mat-hint>
+                <mat-hint align="end">{{ (SecretaryEmailCtrl.value?.length || 0) }}/50</mat-hint>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryEmailCtrl)">
+                <span *ngIf="SecretaryEmailCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryEmailCtrl.hasError('pattern')">ایمیل رو به درستی وارد کنید</span>
+                <span *ngIf="SecretaryEmailCtrl.hasError('maxlength')">حداکثر 50 کاراکتر</span>
+                <span *ngIf="SecretaryEmailCtrl.hasError('server')">{{ SecretaryEmailCtrl.getError('server') }}</span>
+            </div>
+        </div>
+
+        <!-- phone -->
+        <div class="form-item">
+            <mat-form-field appearance="outline" class="phone-field" floatLabel="always">
+                <mat-label>شماره موبایل</mat-label>
+                <span matPrefix dir="ltr" class="phone-label">+98&nbsp;</span>
+                <input class="input" type="tel" maxlength="10" matInput formControlName="phoneNumCtrl" dir="ltr">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryPhoneNumCtrl)">
+                <span *ngIf="SecretaryPhoneNumCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryPhoneNumCtrl.hasError('pattern')">شماره موبایل باید 10 رقم و بدون 0 باشد</span>
+            </div>
+        </div>
+
+        <!-- password -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>رمز عبور</mat-label>
+                <input class="input" [type]="hideSecretaryPassword ? 'password' : 'text'" formControlName="passwordCtrl"
+                    matInput dir="auto">
+                <button mat-icon-button matSuffix tabindex="-1" type="button"
+                    (click)="hideSecretaryPassword = !hideSecretaryPassword">
+                    <mat-icon>{{ hideSecretaryPassword ? 'visibility_off' : 'visibility' }}</mat-icon>
+                </button>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryPasswordCtrl)">
+                <span *ngIf="SecretaryPasswordCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryPasswordCtrl.hasError('minlength')">حداقل 7 کاراکتر</span>
+                <span *ngIf="SecretaryPasswordCtrl.hasError('maxlength')">حداکثر 20 کاراکتر</span>
+                <span *ngIf="SecretaryPasswordCtrl.hasError('server')">{{ SecretaryPasswordCtrl.getError('server')
+                    }}</span>
+            </div>
+        </div>
+
+        <!-- confirm password -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>تکرار رمز عبور</mat-label>
+                <input class="input" [type]="hideSecretaryConfirmPassword ? 'password' : 'text'"
+                    formControlName="confirmPasswordCtrl" matInput dir="auto">
+                <button mat-icon-button matSuffix tabindex="-1" type="button"
+                    (click)="hideSecretaryConfirmPassword = !hideSecretaryConfirmPassword">
+                    <mat-icon>{{ hideSecretaryConfirmPassword ? 'visibility_off' : 'visibility' }}</mat-icon>
+                </button>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(SecretaryConfirmPasswordCtrl)">
+                <span *ngIf="SecretaryConfirmPasswordCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="SecretaryConfirmPasswordCtrl.hasError('minlength')">حداقل 7 کاراکتر</span>
+                <span *ngIf="SecretaryConfirmPasswordCtrl.hasError('maxlength')">حداکثر 20 کاراکتر</span>
+                <span *ngIf="SecretaryConfirmPasswordCtrl.hasError('server')">{{
+                    SecretaryConfirmPasswordCtrl.getError('server') }}</span>
+            </div>
+        </div>
+
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>تاریخ تولد</mat-label>
+                <input matInput placeholder="YYYY-MM-DD" [formControl]="SecretaryDateOfBirthCtrl">
+            </mat-form-field>
+        </div>
+
+        <div class="div-btn-add">
+            <button type="submit" [disabled]="addSecretaryFg.invalid" mat-raised-button color="primary">ثبت نام</button>
+        </div>
+    </div>
+</form>
