+++ client/src/app/components/register/register-secretary/register-secretary.component.scss
@@ -1,0 +1,6 @@
+@use '../shared-form-styles.scss' as *;
+
+:host {
+    display: block;
+    width: 100%;
+}
