+++ client/src/app/components/register/register-student/register-student.component.html
@@ -1,0 +1,122 @@
+<form class="register-form" [formGroup]="addStudentFg" (ngSubmit)="addStudent()" #stuForm="ngForm">
+    <div class="content">
+        <div class="div-title-radio-input">
+            <span class="title">ثبت نام دانشجو</span>
+            <mat-radio-group class="radio-group" aria-label="Select an option" formControlName="genderCtrl">
+                <mat-radio-button class="radio-button" value="male">مرد</mat-radio-button>
+                <mat-radio-button class="radio-button" value="female">زن</mat-radio-button>
+            </mat-radio-group>
+        </div>
+
+        <!-- name -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>نام</mat-label>
+                <input class="input" type="text" formControlName="nameCtrl" matInput dir="auto">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentNameCtrl)">
+                <span *ngIf="StudentNameCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentNameCtrl.hasError('minlength')">خیلی کوتاهه حداقل: 2 حرف</span>
+                <span *ngIf="StudentNameCtrl.hasError('maxlength')">خیلی بلنده حداکثر: 30 حرف</span>
+                <span *ngIf="StudentNameCtrl.hasError('pattern')">فقط حروف فارسی/انگلیسی، فاصله و خط‌تیره مجاز
+                    است.</span>
+            </div>
+        </div>
+
+        <!-- lastName -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>نام خانوادگی</mat-label>
+                <input class="input" type="text" formControlName="lastNameCtrl" matInput dir="auto">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentLastNameCtrl)">
+                <span *ngIf="StudentLastNameCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentLastNameCtrl.hasError('minlength')">خیلی کوتاهه حداقل: 2 حرف</span>
+                <span *ngIf="StudentLastNameCtrl.hasError('maxlength')">خیلی بلنده حداکثر: 30 حرف</span>
+                <span *ngIf="StudentLastNameCtrl.hasError('pattern')">فقط حروف فارسی/انگلیسی، فاصله و خط‌تیره مجاز
+                    است.</span>
+            </div>
+        </div>
+
+        <!-- email -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>ایمیل</mat-label>
+                <input class="input" autoFocusDir type="email" formControlName="emailCtrl" matInput maxlength="50"
+                    dir="auto">
+                <mat-hint>ایمیل خودرا وارد کنید</mat-hint>
+                <mat-hint align="end">{{ (StudentEmailCtrl.value?.length || 0) }}/50</mat-hint>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentEmailCtrl)">
+                <span *ngIf="StudentEmailCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentEmailCtrl.hasError('pattern')">ایمیل رو به درستی وارد کنید</span>
+                <span *ngIf="StudentEmailCtrl.hasError('maxlength')">حداکثر 50 کاراکتر</span>
+                <span *ngIf="StudentEmailCtrl.hasError('server')">{{ StudentEmailCtrl.getError('server') }}</span>
+            </div>
+        </div>
+
+        <!-- phone -->
+        <div class="form-item">
+            <mat-form-field appearance="outline" class="phone-field" floatLabel="always">
+                <mat-label>شماره موبایل</mat-label>
+                <span matPrefix dir="ltr" class="phone-label">+98&nbsp;</span>
+                <input class="input" type="tel" maxlength="10" matInput formControlName="phoneNumCtrl" dir="ltr">
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentPhoneNumCtrl)">
+                <span *ngIf="StudentPhoneNumCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentPhoneNumCtrl.hasError('pattern')">شماره موبایل باید 10 رقم و بدون 0 باشد</span>
+            </div>
+        </div>
+
+        <!-- password -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>رمز عبور</mat-label>
+                <input class="input" [type]="hideStudentPassword ? 'password' : 'text'" formControlName="passwordCtrl"
+                    matInput dir="auto">
+                <button mat-icon-button matSuffix tabindex="-1" type="button"
+                    (click)="hideStudentPassword = !hideStudentPassword">
+                    <mat-icon>{{ hideStudentPassword ? 'visibility_off' : 'visibility' }}</mat-icon>
+                </button>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentPasswordCtrl)">
+                <span *ngIf="StudentPasswordCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentPasswordCtrl.hasError('minlength')">حداقل 7 کاراکتر</span>
+                <span *ngIf="StudentPasswordCtrl.hasError('maxlength')">حداکثر 20 کاراکتر</span>
+                <span *ngIf="StudentPasswordCtrl.hasError('server')">{{ StudentPasswordCtrl.getError('server')
+                    }}</span>
+            </div>
+        </div>
+
+        <!-- confirm password -->
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>تکرار رمز عبور</mat-label>
+                <input class="input" [type]="hideStudentConfirmPassword ? 'password' : 'text'"
+                    formControlName="confirmPasswordCtrl" matInput dir="auto">
+                <button mat-icon-button matSuffix tabindex="-1" type="button"
+                    (click)="hideStudentConfirmPassword = !hideStudentConfirmPassword">
+                    <mat-icon>{{ hideStudentConfirmPassword ? 'visibility_off' : 'visibility' }}</mat-icon>
+                </button>
+            </mat-form-field>
+            <div class="field-error" *ngIf="showErr(StudentConfirmPasswordCtrl)">
+                <span *ngIf="StudentConfirmPasswordCtrl.hasError('required')">فرم رو پرکنید</span>
+                <span *ngIf="StudentConfirmPasswordCtrl.hasError('minlength')">حداقل 7 کاراکتر</span>
+                <span *ngIf="StudentConfirmPasswordCtrl.hasError('maxlength')">حداکثر 20 کاراکتر</span>
+                <span *ngIf="StudentConfirmPasswordCtrl.hasError('server')">{{
+                    StudentConfirmPasswordCtrl.getError('server') }}</span>
+            </div>
+        </div>
+
+        <div class="form-item">
+            <mat-form-field appearance="outline">
+                <mat-label>تاریخ تولد</mat-label>
+                <input matInput placeholder="YYYY-MM-DD" [formControl]="StudentDateOfBirthCtrl">
+            </mat-form-field>
+        </div>
+
+        <div class="div-btn-add">
+            <button type="submit" [disabled]="addStudentFg.invalid" mat-raised-button color="primary">ثبت نام</button>
+        </div>
+    </div>
+</form>
