+++ client/src/app/components/register/register-teacher/register-teacher.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RegisterTeacherComponent } from './register-teacher.component';
+
+describe('RegisterTeacherComponent', () => {
+  let component: RegisterTeacherComponent;
+  let fixture: ComponentFixture<RegisterTeacherComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [RegisterTeacherComponent]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(RegisterTeacherComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
