+++ client/src/app/components/register/shared-form-styles.scss
@@ -1,0 +1,192 @@
+.register-form {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin: 2rem 0;
+
+    .content {
+        width: 90%;
+        display: flex;
+        flex-direction: column;
+        flex-wrap: wrap;
+        gap: 2rem;
+        align-items: center;
+        padding: 1.5rem 1rem;
+
+        border-radius: 10px;
+        background: linear-gradient(145deg, #f0f0f0, #cacaca);
+        box-shadow: 5px 5px 10px #5a5a5a, -5px -5px 10px #ffffff;
+
+        .div-title-radio-input {
+            flex: 0 0 100%;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            width: 100%;
+            margin-top: .5rem;
+            gap: 1rem;
+
+            .title {
+                font-size: 20px;
+                color: #000;
+                font-weight: 800;
+            }
+
+            .radio-group {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                gap: 1rem;
+
+                .radio-button {
+                    height: 60px;
+                    width: 100px;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
+                    border-radius: 10px;
+                    background: linear-gradient(145deg, #f0f0f0, #cacaca);
+                    box-shadow: 5px 5px 10px #5a5a5a, -5px -5px 10px #ffffff;
+                }
+            }
+        }
+
+        .form-item {
+            width: 90%;
+            max-width: 700px;
+            display: flex;
+            flex-direction: column;
+            align-items: stretch;
+            min-width: 0;
+
+            mat-form-field {
+                display: block;
+                width: 100%;
+
+                .input {
+                    padding: 10px;
+                    height: 54px;
+                    width: 100%;
+                    border: none;
+                    border-radius: 10px;
+                    background: transparent;
+                    box-shadow: none;
+                }
+            }
+
+            .field-error {
+                margin-top: 6px;
+                padding: 0 4px;
+                font-size: 12px;
+                color: #d32f2f;
+                line-height: 1.6;
+                display: flex;
+                flex-direction: column;
+                gap: 2px;
+                width: 100%;
+            }
+        }
+
+        .div-btn-add {
+            flex: 0 0 100%;
+            width: 100%;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            margin: .5rem 0 1rem;
+
+            button {
+                border-radius: 10px;
+                border: none;
+                height: 40px;
+                width: 120px;
+            }
+        }
+    }
+}
+
+@media (min-width: 768px) {
+    .register-form .content {
+        width: 90%;
+        flex-direction: row;
+        justify-content: center;
+        align-items: stretch;
+    }
+
+    .register-form .content .form-item {
+        flex: 1 1 45%;
+        max-width: 45%;
+        width: auto;
+    }
+}
+
+@media (min-width: 1200px) {
+    .register-form .content {
+        width: 60%;
+        justify-content: space-between;
+    }
+
+    .register-form .content .form-item {
+        flex: 1 1 30%;
+        max-width: 30%;
+    }
+}
+
+::ng-deep mat-form-field {
+    height: auto !important;
+    margin-bottom: 10px;
+}
+
+::ng-deep .mat-mdc-text-field-wrapper {
+    border-radius: 12px;
+    background: #e0e0e0;
+    box-shadow: inset 5px 5px 10px #cecece, inset -5px -5px 10px #f2f2f2;
+    padding: 0;
+    overflow: visible;
+}
+
+::ng-deep .mat-mdc-form-field-appearance-outline .mdc-notched-outline {
+    display: none;
+}
+
+::ng-deep .mat-mdc-form-field-infix {
+    text-align: right;
+    direction: rtl;
+    padding: 12px 14px !important;
+    min-height: 54px;
+    width: 100%;
+    box-sizing: border-box;
+}
+
+::ng-deep .mat-mdc-form-field-subscript-wrapper {
+    position: relative;
+    bottom: 0;
+    margin-top: 6px;
+    padding: 0 12px;
+    height: auto;
+}
+
+::ng-deep mat-form-field.phone-field .mat-mdc-form-field-prefix {
+    direction: ltr;
+    display: flex;
+    align-items: center;
+    height: 100%;
+    padding-inline: 4px;
+    font-weight: 700;
+    opacity: .9;
+}
+
+::ng-deep mat-form-field.phone-field input.mat-mdc-input-element {
+    line-height: 1.5;
+    text-align: left;
+}
+
+::ng-deep mat-form-field.phone-field .mat-mdc-text-field-wrapper {
+    padding: 0 12px;
+}
+
+::ng-deep mat-form-field.phone-field .phone-label {
+    margin-left: 10px;
+}
